refactor(browser): group edge/browser into vtd-clients/ + transport/ subdirs
The 5-cluster browser-mode integration grew webapp/src/shell/edge/browser to 19
children, tripping the tier-0 directory-fanout gate (cap 15). Split by concern:
vtd-clients/ (vtdRpc + the vtd* RPC wrappers + their tests) and transport/
(graphEventStream, reconnectPolicy, browserTerminal + tests). Top level keeps the
adapter entry (browserRuntime), bootstrap (installBrowserRuntime), config, mic.
Pure file moves + import-path updates — zero logic change. tsc clean, 44 browser
tests pass.
