+++ webapp/src/shell/edge/browser/transport/graphEventStream.ts
