+++ webapp/src/shell/edge/browser/vtd-clients/vtdImageClient.ts
